requirements.txt
langchain>=0.2.0
chromadb>=0.4.24
google-generativeai>=0.4.0
tqdm>=4.66
langchain-community>=0.0.26
python-dotenv>=1.0.1  
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9